HelloMoonFragment AudioPlayer bugbfix.
AudioPlayer object creation was moved to mPlayer var declaration.

The problem is that the object was created in onCreateView() and since
that method is called every time the view for the fragment needs to be
created (and inflated) it would cause the creation of multiple
AudioPlayers and cause bugs.

Retaining a Fragment doesn't retain its view, it gets recreated every time
the Fragmen is destroyed (or Retained, in this case).